+++ src/pages/Router.jsx
@@ -29,7 +29,6 @@ import PostWrite from "./community/PostWrite";
 import ComparePrograms from "./program/ComparePrograms";
 import Sports from "./sports/Sports";
 import PostEdit from "./community/PostEdit";
-import SportsInfo from "./sports/SportsInfo";
 
 export default function Router() {
   return (
@@ -62,7 +61,6 @@ export default function Router() {
 
         {/* Sports */}
         <Route path="sports" element={<Sports />} />
-        <Route path="sports/:sport" element={<SportsInfo />} />
 
         {/* Matching */}
         <Route path="matching" element={<Matching />} />
